moved inference loop under compute
